Updating API & Editor
Updated API to reference the clone of the original div for the editor not the original itself. This allows for a user to enter their own code and test it.

I need to code in more status ids to account for any code errors a user might try to compile.